requirements.txt
torch==2.7.0
torchvision==0.22.0
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=0.27.0
numpy>=1.26.0
pillow>=10.0.0
matplotlib>=3.8.0
tqdm>=4.66.0